Move event log format utils to common location
