+++ indico/modules/events/operations.py
@@ -22,6 +22,7 @@ from indico.modules.events.management.settings import privacy_settings
 from indico.modules.events.models.events import EventType
 from indico.modules.events.models.labels import EventLabel
 from indico.modules.events.models.references import ReferenceType
+from indico.modules.events.util import format_log_person, format_log_ref, split_log_location_changes
 from indico.modules.logs.models.entries import CategoryLogRealm, LogKind
 from indico.modules.logs.util import make_diff_log
 from indico.util.i18n import orig_string
@@ -235,11 +236,11 @@ def _log_event_update(event, changes, extra_log_fields, visible_person_link_chan
         'keywords': 'Keywords',
         'references': {
             'title': 'External IDs',
-            'convert': lambda changes: [list(map(_format_ref, refs)) for refs in changes]
+            'convert': lambda changes: [list(map(format_log_ref, refs)) for refs in changes]
         },
         'person_links': {
             'title': 'Speakers' if event.type_ == EventType.lecture else 'Chairpersons',
-            'convert': lambda changes: [list(map(_format_person, persons)) for persons in changes]
+            'convert': lambda changes: [list(map(format_log_person, persons)) for persons in changes]
         },
         'start_dt': 'Start date',
         'end_dt': 'End date',
@@ -259,7 +260,7 @@ def _log_event_update(event, changes, extra_log_fields, visible_person_link_chan
         'enforce_locale': 'Enforce language',
         **(extra_log_fields or {})
     }
-    _split_location_changes(changes)
+    split_log_location_changes(changes)
     if not visible_person_link_changes:
         # Don't log a person link change with no visible changes (changes
         # on an existing link or reordering). It would look quite weird in
@@ -278,42 +279,6 @@ def _log_event_update(event, changes, extra_log_fields, visible_person_link_chan
             what = 'Data'
         event.log(EventLogRealm.management, LogKind.change, 'Event', f'{what} updated', session.user,
                   data={'Changes': make_diff_log(changes, log_fields)})
-
-
-def _format_ref(ref):
-    return f'{ref.reference_type.name}:{ref.value}'
-
-
-def _get_venue_room_name(data):
-    venue_name = data['venue'].name if data.get('venue') else data.get('venue_name', '')
-    room_name = data['room'].full_name if data.get('room') else data.get('room_name', '')
-    return venue_name, room_name
-
-
-def _format_location(data):
-    venue_name = data[0]
-    room_name = data[1]
-    if venue_name and room_name:
-        return f'{venue_name}: {room_name}'
-    elif venue_name or room_name:
-        return venue_name or room_name
-    else:
-        return None
-
-
-def _split_location_changes(changes):
-    location_changes = changes.pop('location_data', None)
-    if location_changes is None:
-        return
-    if location_changes[0]['address'] != location_changes[1]['address']:
-        changes['address'] = (location_changes[0]['address'], location_changes[1]['address'])
-    venue_room_changes = (_get_venue_room_name(location_changes[0]), _get_venue_room_name(location_changes[1]))
-    if venue_room_changes[0] != venue_room_changes[1]:
-        changes['venue_room'] = list(map(_format_location, venue_room_changes))
-
-
-def _format_person(data):
-    return f'{data.full_name} <{data.email}>' if data.email else data.full_name
 
 
 @make_interceptable
